+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Made sure that only login managers can see logins other than themselves, even when they have the token.
 - Added a language assignment to logins.
 - Added the basics for a set of strings that can be associated with token IDs.
+- When a user logs in, the last_access column of the login record is touched.
 
 **1.0.0.2019** *(May 10, 2018)*